* Fixed make install (irrlicht now does not try to install) * Updated SysDefs to work better on 64bit systems, please report problems on 32bit systems * Various x64 related fixes * Updated movementflags to recent mangos version * ... and fixed movement packets. not sure if that is the best way...

+++ src/shared/SysDefs.h
@@ -75,11 +75,11 @@
     #define I64FMTD "%llu"
     #define SI64FMTD "%lld"
     typedef long long int64;
-    typedef long int32;
+    typedef int int32;
     typedef short int16;
     typedef char int8;
     typedef unsigned long long uint64;
-    typedef unsigned long uint32;
+    typedef unsigned int uint32;
     typedef unsigned short uint16;
     typedef unsigned char uint8;
     typedef unsigned short WORD;
